+++ src/runtime/server/routes/messages.ts
@@ -1,7 +1,7 @@
 import { deepCopy } from '@intlify/shared'
 import { defineCachedEventHandler, defineCachedFunction } from 'nitropack/runtime'
 import { createError, defineEventHandler, getRouterParam } from 'h3'
-import { useI18nContext } from '../context'
+import { initializeI18nContext, tryUseI18nContext, useI18nContext } from '../context'
 import { getMergedMessages } from '../utils/messages'
 
 import type { H3Event } from 'h3'
@@ -31,10 +31,12 @@ const _cachedMessageLoader = defineCachedFunction(_messagesHandler, {
   name: 'i18n:messages-internal',
   maxAge: !__I18N_CACHE__ ? -1 : 60 * 60 * 24,
   getKey: event => [getRouterParam(event, 'locale') ?? 'null', getRouterParam(event, 'hash') ?? 'null'].join('-'),
-  shouldBypassCache(event) {
+  async shouldBypassCache(event) {
     const locale = getRouterParam(event, 'locale')
     if (locale == null) { return false }
-    return !useI18nContext(event).localeConfigs?.[locale]?.cacheable
+    // prerendering may require initializing context
+    const ctx = tryUseI18nContext(event) || await initializeI18nContext(event)
+    return !ctx.localeConfigs?.[locale]?.cacheable
   },
 })
